fix: shape removal, proactive type safety, and alignment readback
- Add shape/picture removal support in Excel Remove handler
- Add shape/picture query support in Excel Query handler
- Fix 8 non-standard types in Format dict (byte/uint → int) preventing
  JSON serialization crashes
- Add textRotation, indent, shrinkToFit readback in cell Get
- Fix PowerPoint shadow=true and glow=true crash in Add/Set
- Add comprehensive tests for all fixes

+++ src/officecli/Core/ChartReader.cs
@@ -47,7 +47,7 @@ internal static partial class ChartHelper
 
         // Chart style
         var style = chart.Parent?.GetFirstChild<C.Style>();
-        if (style?.Val?.HasValue == true) node.Format["style"] = style.Val.Value;
+        if (style?.Val?.HasValue == true) node.Format["style"] = (int)style.Val.Value;
 
         // Plot area fill (plotArea uses C.ShapeProperties, not C.ChartShapeProperties)
         var plotSpPr = plotArea.GetFirstChild<C.ShapeProperties>();
